+++ src/main/java/com/starfish_studios/naturalist/server/entity/ai/goal/CloseMeleeAttackGoal.java
@@ -1,6 +1,7 @@
 package com.starfish_studios.naturalist.server.entity.ai.goal;
 
 import net.minecraft.util.Mth;
+import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.PathfinderMob;
 import net.minecraft.world.entity.ai.goal.MeleeAttackGoal;
@@ -16,6 +17,7 @@ public class CloseMeleeAttackGoal extends MeleeAttackGoal {
         double reach = Mth.square(this.mob.getBbWidth() * 1.2f);
         if (this.mob.distanceToSqr(target) <= reach && this.isTimeToAttack()) {
             this.resetAttackCooldown();
+            this.mob.swing(InteractionHand.MAIN_HAND);
             this.mob.doHurtTarget(target);
         }
     }