+++ src/main/java/com/starfish_studios/naturalist/server/entity/mob/Alligator.java
@@ -254,9 +254,8 @@ public class Alligator extends NaturalistAnimal implements NaturalistGeoEntity, 
     }
 
     private <E extends Alligator> PlayState attackPredicate(final AnimationState<E> event) {
-        if (this.swinging && event.getController().getAnimationState().equals(AnimationController.State.STOPPED)) {
+        if (this.swinging) {
             event.getController().forceAnimationReset();
-
             event.getController().setAnimation(BITE);
             this.swinging = false;
         }